refactor: modularize rginx-config stage 3
